requirements.txt
influxdb
python-telegram-bot
configparser
requests
